requirements.txt
PyQt6>=6.5.0
httpx>=0.25.0
openai>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
allure-pytest>=2.13.0
pyinstaller>=6.0.0
